backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdfplumber==0.11.0
python-docx==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
